pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "andx"
version = "0.1.0"
description = "AND-2025: Anchor Neighbourhood Discovery with Automatic K"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Shahid Ahamed Hasib"},
  {name = "Imene MOULAY OMAR"}
]
dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "numpy>=1.26",
  "scipy>=1.11",
  "tqdm>=4.66",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "tensorboard>=2.14",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "einops>=0.7",
  "rich>=13.7",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

